Scale and room changes, camera implemented
From the top:

Collider now sets scale but not with a global variable

Removed renderersetscale. Drawmanager now draws the sprites larger by a
factor of m_scale

Engine added a escape feature

Gamestate removed single wall. Rooms now loads with scale. Sprites now
update with deltatime. Room draw loops are now correct and updated with
scale. Drawmanager now always draw the player in the middle and
everything else offset to create a camera.

Item gets a virtual destructor.

Player gets more speed to accommodate the scale. Setposition now takes
floats.

Rooms for loop fixed and inclusion of scale.

Sprite gets a virtual destructor.

Spriteanimation sprites are now updated per second instead of per frame.

Size of colliders are now set in room::load.

+++ Player.h
@@ -17,7 +17,7 @@ public:
 	float GetY();
 	bool IsVisible();
 	EEntityType GetType();
-	void SetPosition(int x, int y);
+	void SetPosition(float x, float y);
 	float GetDirectionX();
 	float GetDirectionY();
